requirements.txt
# Deploy
streamlit==1.35.0
lmdeploy==0.4.0
modelscope==1.14.0
opencv-python==4.9.0.80

# RAG
BCEmbedding==0.1.5
langchain==0.2.0
langchain-community==0.2.0
loguru==0.7.2
textract==1.6.5
faiss-gpu==1.7.2

# Train
xtuner[deepspeed]==0.1.19

# TTS 1.0 Text to Speech
# pytorch_wavelets==1.3.0
# tensorboardX==2.6.2.2
# scipy==1.12.0

# -f https://modelscope.oss-cn-beijing.aliyuncs.com/releases/repo.html
# kantts

# TTS 2.0
LangSegment==0.3.3
librosa==0.9.2
ffmpeg-python==0.2.0
pytorch-lightning==2.2.5
cn2an==0.5.22
pypinyin==0.51.0
jieba-fast==0.53
wordsegment==1.3.1
g2p-en==2.1.0
